+++ problem-solver/cxx/inferenceModule/generator/SolutionTreeGenerator.cpp
@@ -20,11 +20,12 @@ SolutionTreeGenerator::SolutionTreeGenerator(ScMemoryContext * ms_context)
       : ms_context(ms_context)
 {
   solution = ms_context->CreateNode(ScType::NodeConst);
+  ms_context->CreateEdge(ScType::EdgeAccessConstPosPerm, InferenceKeynodes::concept_solution, solution);
 }
 
-bool SolutionTreeGenerator::addNode(ScAddr const & formula, ScTemplateParams const & templateParams)
+bool SolutionTreeGenerator::addNode(ScAddr const & formula, ScTemplateParams const & templateParams, std::set<std::string> const & varNames)
 {
-  ScAddr newSolutionNode = createSolutionNode(formula, templateParams);
+  ScAddr newSolutionNode = createSolutionNode(formula, templateParams, varNames);
   bool result = newSolutionNode.IsValid();
   if (result)
   {
@@ -59,11 +60,31 @@ bool SolutionTreeGenerator::addNode(ScAddr const & formula, ScTemplateParams con
   return result;
 }
 
-ScAddr SolutionTreeGenerator::createSolutionNode(ScAddr const & formula, ScTemplateParams const & templateParams)
+ScAddr SolutionTreeGenerator::createSolutionNode(ScAddr const & formula, ScTemplateParams const & templateParams, std::set<std::string> const & varNames)
 {
-  ScAddr solutionNode = ms_context->CreateNode(ScType::NodeConst);
-  // TODO: Add params to solution node
+  ScAddr const & solutionNode = ms_context->CreateNode(ScType::NodeConst);
   GenerationUtils::generateRelationBetween(ms_context, solutionNode, formula, CoreKeynodes::rrel_1);
+  ScAddr const & replacementsNode = ms_context->CreateNode(ScType::NodeConst);
+  GenerationUtils::generateRelationBetween(ms_context, solutionNode, replacementsNode, CoreKeynodes::rrel_2);
+  for (std::string const & varName : varNames)
+  {
+    ScAddr replacement;
+    templateParams.Get(varName, replacement);
+    if (replacement.IsValid())
+    {
+      ScAddr const & pair = ms_context->CreateNode(ScType::NodeConst);
+      ms_context->CreateEdge(ScType::EdgeAccessConstPosPerm, replacementsNode, pair);
+      ScAddr const & varNode = ms_context->HelperFindBySystemIdtf(varName);
+      GenerationUtils::generateRelationBetween(ms_context, pair, replacement, CoreKeynodes::rrel_1);
+      GenerationUtils::generateRelationBetween(ms_context, pair, varNode, CoreKeynodes::rrel_2);
+      ms_context->CreateEdge(ScType::EdgeAccessConstPosTemp, varNode, replacement);
+    }
+    else
+      SC_THROW_EXCEPTION(utils::ScException,
+                         "SolutionTreeGenerator: formula " << ms_context->HelperGetSystemIdtf(formula) << " has var " <<
+                         varName << " but scTemplateParams don't have replacement for this var");
+  }
+
   return solutionNode;
 }
 
@@ -71,7 +92,6 @@ ScAddr SolutionTreeGenerator::createSolution(ScAddr const & outputStructure, boo
 {
   ScType arcType = targetAchieved ? ScType::EdgeAccessConstPosPerm : ScType::EdgeAccessConstNegPerm;
   ms_context->CreateEdge(arcType, InferenceKeynodes::concept_success_solution, solution);
-  ms_context->CreateEdge(ScType::EdgeAccessConstPosPerm, InferenceKeynodes::concept_solution, solution);
   GenerationUtils::generateRelationBetween(ms_context, solution, outputStructure, InferenceKeynodes::nrel_output_structure);
 
   return solution;