+++ problem-solver/cxx/inferenceModule/utils/ReplacementsUtils.hpp
@@ -25,9 +25,9 @@ public:
   static vector<ScTemplateParams> getReplacementsToScTemplateParams(Replacements const & replacements);
   static size_t getColumnsAmount(Replacements const & replacements);
   static size_t getRowsAmount(Replacements const & replacements);
+  static set<string> getKeySet(Replacements const & map);
 
 private:
-  static set<string> getKeySet(Replacements const & map);
   static set<string> getCommonKeys(set<string> const & first, set<string> const & second);
   static Replacements copyReplacements(Replacements const & replacements);
 };